+++ src/kvstore.h
@@ -8,10 +8,10 @@
 #include <iostream>
 #include <unistd.h>
 #include <cstdlib>
+#include "dataframe.h"
 
 using namespace std;
 
-class DataFrame;
 
 /**
  * Represents a Key Value Store
@@ -63,8 +63,8 @@ public:
         cout << " we put" << endl;
 
         // if new key add and increment size
-        *this->keys[size] = *key;
-        *this->dfs[size] = *df;
+        this->keys[size] = key;
+        this->dfs[size] = df;
 
 //        cout << "dfs[0]- " << dfs[0] << endl;
